+++ packages/core/src/rules/unnecessary_chaining.ts
