+++ packages/core/test/rules/unnecessary_chaining.ts
